The Green Palace temporary fix
This pull request replaces the Specific Tile condition for a free Great Farmer. This is not meant to be a permanent fix. Another simple solution could be to have a Great Botanist unit that can spawn only from the Green Palace wonder and can only make a Agricultural Farm building.

I made this because of a bug I encountered, you can see it here: https://github.com/SpacedOutChicken/Deciv-2/issues/16